refactor: 패키지 구조를 개선하여 도메인 모델을 명확히 구분
