+++ src/main/java/car/domain/model/Name.java
@@ -1,4 +1,4 @@
-package car.domain;
+package car.domain.model;
 
 public class Name {
