+++ src/main/java/car/domain/strategy/MovingStrategy.java
@@ -1,4 +1,4 @@
-package car.domain;
+package car.domain.strategy;
 
 @FunctionalInterface
 public interface MovingStrategy {
